renamed CacheCleaner to CacheItemSelector
